cleancode
# message auto-generated for no-merge-commit merge:
merge cleancode_master into master

cleancode整改

Created-by: msprofDeveloper

Match-id-c90e77458294de7290c49c2625f12b6a05e59796

+++ analysis/csrc/domain/data_process/ai_task/metric_processor.cpp
@@ -379,7 +379,7 @@ bool MetricProcessor::CheckAndGetMemoryBoundRelatedDataIndex(const std::vector<s
         mac_exe_ratio, vec_exe_ratio, mte2_exe_ratio};
     int count = 0;
     neededDataIndex = {0, 0, 0};
-    const int neededDataSize = 3;
+    const uint16_t neededDataSize = 3;
     for (unsigned index = 0; index < neededData.size(); ++index) {
         for (unsigned headersIndex = 0; headersIndex < headers.size(); ++headersIndex) {
             if (neededData[index] == headers[headersIndex]) {
